Game dir detection fixed, workaround for 'open folder' bug
Fixed detection of game dir for the Telltale versions of The Walking
Dead Season 2 Episode 1 and The Wolf Among Us. Previously it only
auto-detected it for the Steam versions.

Added workaround for game detection if someone chooses 'open folder'
rather than choosing a specific game. This bug only affects The Walking
Dead Season 2 Episode 1 and The Wolf Among Us.